requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
faiss-cpu>=1.7.4
cohere>=5.5.0
boto3>=1.34.0
pydantic>=2.8.0
python-dotenv>=1.0.1
orjson>=3.10.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

